+++ src/utils/functions/sort/sort-by-property.ts
@@ -2,11 +2,13 @@ import type { addPrefixToObject } from '../../types/add-prefix-to-object';
 import type { PropOf } from '../../types/prop-of';
 import type { PropertyOnly } from '../../types/property-only';
 import type { SortFunction } from '../../types/sort-function';
+import type { SortOrder } from '../../types/sort-order';
+import { sortValues } from './sort-values';
 
 export function sortByProperty<T extends object>(
   property: PropOf<T> | keyof addPrefixToObject<PropertyOnly<T>, '-'>
 ): SortFunction<T> {
-  let sortOrder: 1 | -1 = 1;
+  let sortOrder: SortOrder = 1;
   let prop = property as string;
 
   if (prop.startsWith('-')) {
@@ -20,13 +22,6 @@ export function sortByProperty<T extends object>(
     const valueA = a[key];
     const valueB = b[key];
 
-    if (valueA == null && valueB == null) return 0;
-    if (valueA == null) return 1 * sortOrder;
-    if (valueB == null) return -1 * sortOrder;
-
-    if (valueA < valueB) return -1 * sortOrder;
-    if (valueA > valueB) return 1 * sortOrder;
-
-    return 0;
+    return sortValues(valueA, valueB, sortOrder);
   };
 }